Fixed a service collection infinite recursion bug.
This only would have been seen in certain corner cases.
When auto selecting a service collection tag, it would try to use a tag
that it got from a class variable. That class only got included once the
resource it was trying to collect got defined. This accident happened
because I was really looking at the wrong variable for collection.