Make hacking on FreeIPA faster.
If you happen to make a custom base box that already has the ipa-server
package installed, then vagrant will use this box. This speeds up the
RPM download and installation process. Skipping this will allow you to
iterate more quickly. The downside is you have to regenerate the box if
you want it to have the latest version of FreeIPA.