Add sorted uuidgen -t based ordering for joined fqdn's.
This ensures that if new peers join the cluster after an initial built,
then then ordered list doesn't flip-flop. That is, they should always
get appended to the end of the list. Keep in mind that what the peering
algorithm does with them after that is a different issue.

Thanks to Pádraig Brady for giving me a sorted uuidgen -t algorithm.