Avoid unnecessary errors from ipa-replica-manage connect.
Since the topology can be known algorithmically before the puppet-ssh
module has had a chance to transfer the replica-info-*.gpg file, then
the ipa-replica-manage connect command can try and connect a peer before
the peer has even had a chance to ipa-replica-install first. This check
ensures that there is an ipa server up and listening first so that we
don't show unnecessary errors.

Of particular note is that I believe that I am somewhat mis-using the
ipa-replica-conncheck command to do this verification. It is actually
being used (AFAICT) to do exactly what it was designed to do, but I am
using it in an undocumented way, because I am expecting it to work
against a real ipa replica, instead of the fake listener that you would
run on your peer server with the ipa-replica-conncheck tool.

This error actually showed up in a strange way on 3.0.0, see:
https://fedorahosted.org/freeipa/ticket/3105

It should show up more intelligibly in 3.0.1, although it should
probably never be seen with this patch (unless there is a real problem).